fix: cleanup
- error handling in download pdf, csv
- cleanups code
